Move JobMetadataInjector out of decorators
